Correção de parte do modelo de observaçao dos sonares. Bug no RobotDataGatherer. Mais testes unitários.

+++ trunk/Sauron/Sonar/CustomTypes.h
@@ -2,6 +2,8 @@
 
 namespace sauron
 {
-	typedef int reading_t;
+	// leitura de um sonar, em cm
+	typedef double reading_t;
+	// posição global em X (cm), Y (cm) ou th (rad)
 	typedef double pose_t;
 }
